chore: update import path after dir restructure
