+++ src/app/(app)/components/LogoutButton.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/Button';
 import { signOut } from '@/services/signOut';
 import { LogOut } from 'lucide-react';
 import { useRouter } from 'next/navigation';
@@ -26,7 +26,7 @@ export function LogoutButton() {
     <Button
       onClick={handleSignOut}
       disabled={isLoading}
-      className="gap-2 py-0 px-0 font-normal"
+      className="gap-2 py-0 px-0 font-normal disabled:opacity-45 disabled:border-none"
     >
       <span>Sair</span> <LogOut size={14} />
     </Button>
