+++ src/app/(auth)/forget-password/components/ForgetPasswordForm.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/Button';
 import { submitForgetPassword } from '@/services/submitForgetPassword';
 import { useRouter } from 'next/navigation';
 import { FormEvent, useState } from 'react';
