+++ src/app/(auth)/waiting-code/components/WaitingCodeForm.tsx
@@ -1,6 +1,6 @@
 'use client';
+import { Button } from '@/components/Button';
 import { InputOTP, InputOTPGroup, InputOTPSlot } from '@/components/InputOPT';
-import { Button } from '@/components/ui/button';
 import { submitWaitingCode } from '@/services/submitWaitingCode';
 import Link from 'next/link';
 import { useRouter, useSearchParams } from 'next/navigation';
